UpCommand was missing equals and hashcode. Refactored Compiler, Executor and javac.Utils. Wrote a timeout test for javac package.

+++ src/main/java/javac/Utils.java
@@ -4,26 +4,49 @@ import dao.Privacy;
 
 import java.io.*;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 class Utils {
 
     private Utils() {} // prevent instantiation
 
-    static void write(ClassFile classFile) throws IOException {
-        Files.write(Paths.get(classFile.getClassName() + ".class"), classFile.getByteCode());
+    static Path write(ClassFile classFile) {
+        if (classFile == null) return null;
+        try {
+            return Files.write(Paths.get(classFile.getClassName() + ".class"), classFile.getByteCode());
+        } catch (IOException e) {
+            return null;
+        }
     }
 
-    static void write(JavaFile javaFile) throws IOException {
-        Files.write(Paths.get(javaFile.getClassName()  + ".java"), javaFile.getSource().getBytes());
+    static Path write(JavaFile javaFile) {
+        if (javaFile == null) return null;
+        try {
+            return Files.write(Paths.get(javaFile.getClassName()  + ".java"), javaFile.getSource().getBytes());
+        } catch (IOException e) {
+            return null;
+        }
     }
 
-    static void delete(JavaFile javaFile) throws IOException {
-        Files.delete(Paths.get(javaFile.getClassName() + ".java"));
+    static boolean delete(JavaFile javaFile) {
+        if (javaFile == null) return false;
+        try {
+            Files.delete(Paths.get(javaFile.getClassName() + ".java"));
+            return true;
+        } catch (IOException e) {
+            return false;
+        }
     }
 
-    static void delete(ClassFile classFile) throws IOException {
-        Files.delete(Paths.get(classFile.getClassName() + ".class"));
+    static boolean delete(ClassFile classFile) {
+        if (classFile == null) return false;
+        try {
+            Files.delete(Paths.get(classFile.getClassName() + ".class"));
+            return true;
+        } catch (IOException e) {
+            return false;
+        }
     }
 
     static ClassFile readClassFile(JavaFile javaFile) {
@@ -35,18 +58,17 @@ class Utils {
         }
     }
 
-    static void write(ClassFile classFile, Privacy privacy, Long id) {
-
-    }
-
-    static String getLines(InputStream is) throws IOException {
+    static String getLines(InputStream is) {
         String line;
         StringBuilder sb = new StringBuilder();
         BufferedReader in = new BufferedReader(new InputStreamReader(is)); //todo add charset
-        while ((line = in.readLine()) != null) {
-            sb.append(line);
-            sb.append(System.getProperty("line.separator"));
+
+        try {
+            while ((line = in.readLine()) != null) sb.append(line).append(System.getProperty("line.separator"));
+        } catch (IOException e) {
+            return "";
         }
+
         return sb.toString();
     }
 }
